Breakdown the Server into smaller pieces.
