+++ ssh_server/configuration.go
@@ -1,0 +1,26 @@
+package main
+
+import (
+	"fmt"
+	"os"
+
+	"github.com/robinje/multi-user-dungeon/core"
+	"gopkg.in/yaml.v3"
+)
+
+// loadConfiguration reads the configuration file and unmarshals it into a Configuration struct.
+func loadConfiguration(configFile string) (core.Configuration, error) {
+	var config core.Configuration
+
+	data, err := os.ReadFile(configFile)
+	if err != nil {
+		return config, fmt.Errorf("error reading config file: %w", err)
+	}
+
+	err = yaml.Unmarshal(data, &config)
+	if err != nil {
+		return config, fmt.Errorf("error unmarshalling config: %w", err)
+	}
+
+	return config, nil
+}
